+++ src/main/java/racingcar/configuration/AppConfig.java
@@ -1,8 +1,8 @@
 package racingcar.configuration;
 
-import racingcar.NameSeparator;
-import racingcar.Printer;
-import racingcar.Reader;
+import racingcar.racingapplication.NameSeparator;
+import racingcar.racingapplication.Printer;
+import racingcar.racingapplication.Reader;
 import racingcar.nameseparator.OriginSeparator;
 import racingcar.printer.PrintRacingResult;
 import racingcar.reader.ConsoleReader;
