+++ src/main/java/racingcar/racingapplication/Reader.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.racingapplication;
 
 public interface Reader {
